samples: portable uptime format — %lld with explicit (long long) cast
Justin hit -Wformat on wifi_init's native_sim build: k_uptime_get()
returns int64_t, which is plain long on the LP64 native_sim host (so
%lld warns there) but long long on the ILP32 hardware targets (so a bare
%ld fixes native_sim while mis-formatting a 64-bit vararg on ESP32-C6/
nRF91 and moving the warning to those builds). The cast form is correct
and silent on both. Applied to all five samples' report_uptime()
(identical line each). Verified: wifi_init native_sim pristine rebuild,
zero format warnings, binary produced.

+++ samples/asset_tracker/src/shadow.c
@@ -75,7 +75,7 @@ static void report_metric(const char *key, const char *val) {
 static void report_uptime(void) {
   char uptime_s[16];
 
-  snprintk(uptime_s, sizeof(uptime_s), "%lld", k_uptime_get() / 1000);
+  snprintk(uptime_s, sizeof(uptime_s), "%lld", (long long)(k_uptime_get() / 1000));
   report_metric("uptime_s", uptime_s);
 }
